fix: override whatwg-url version to prevent punycode warning
The openai package transitively depends on an outdated version of
whatwg-url, which causes nodejs to print a deprecation warning for
punycode when running the vision tool. This will be fixed when openai
resolves https://github.com/openai/openai-node/pull/392. Until then,
override the version of whatwg-url in order to prevent the deprecation
warning.

See https://github.com/openai/openai-node/issues/527#issuecomment-1823471697
for details on this stopgap.

Signed-off-by: Nick Hale <4175918+njhale@users.noreply.github.com>